Move everything out of lib.rs into their own modules
