+++ src/krustlet.rs
@@ -1,0 +1,29 @@
+use k8s_openapi::api::core::v1::Toleration;
+
+/// Creates a vector of tolerations we need to work with the Krustlet.
+/// Usually these would be added to a Pod so it can be scheduled on a Krustlet.
+pub fn create_tolerations() -> Vec<Toleration> {
+    vec![
+        Toleration {
+            effect: Some(String::from("NoExecute")),
+            key: Some(String::from("kubernetes.io/arch")),
+            operator: Some(String::from("Equal")),
+            toleration_seconds: None,
+            value: Some(String::from("stackable-linux")),
+        },
+        Toleration {
+            effect: Some(String::from("NoSchedule")),
+            key: Some(String::from("kubernetes.io/arch")),
+            operator: Some(String::from("Equal")),
+            toleration_seconds: None,
+            value: Some(String::from("stackable-linux")),
+        },
+        Toleration {
+            effect: Some(String::from("NoSchedule")),
+            key: Some(String::from("node.kubernetes.io/network-unavailable")),
+            operator: Some(String::from("Exists")),
+            toleration_seconds: None,
+            value: None,
+        },
+    ]
+}
